PRF_PROG1_20232_HernandezJuanPablo_EspitiaAlejandra_GuerreroValentina_PenicheGianfranco/src/co/edu/unbosque/test/GestionApostadorTest.java
package co.edu.unbosque.test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;
import static org.junit.Assert.assertFalse;
import org.junit.Before;
import org.junit.Test;

import co.edu.unbosque.model.GestionApostador;

/**
 * Clase de pruebas para la clase GestionApostador.
 */
public class GestionApostadorTest {

    private GestionApostador gestionApostador;

    @Before
    public void setUp() {
        gestionApostador = new GestionApostador();
    }

    /**
     * Prueba para verificar la creación correcta de un apostador.
     */
    @Test
    public void testCrearApostador() {
        gestionApostador.crearApostador("Juan", 123456789, "Sede1", "Calle 123", 1234567890);
        // Puedes agregar verificaciones adicionales aquí según la lógica de tu aplicación.
        // assertTrue("Condición esperada", condicionActual);
    }

    /**
     * Prueba para verificar la modificación correcta de un apostador.
     */
    @Test
    public void testModificarApostador() {
        gestionApostador.crearApostador("Juan", 123456789, "Sede1", "Calle 123", 1234567890);
        gestionApostador.modificarApostador("Juan", "NuevoJuan", 987654321, "Sede2", "Calle 456", 987654321);
        // Puedes agregar verificaciones adicionales aquí según la lógica de tu aplicación.
        // assertTrue("Condición esperada", condicionActual);
    }

    /**
     * Prueba para verificar la lectura correcta de todos los apostadores.
     */
    @Test
    public void testLeerApostadores() {
        gestionApostador.crearApostador("Juan", 123456789, "Sede1", "Calle 123", 1234567890);
        String resultado = gestionApostador.leerApostadores();
        // Puedes verificar el resultado esperado o realizar otras comprobaciones.
        // assertEquals("Resultado Esperado", resultado);
    }

    /**
     * Prueba para verificar la eliminación correcta de un apostador.
     */
    @Test
    public void testEliminarApostador() {
        gestionApostador.crearApostador("Juan", 123456789, "Sede1", "Calle 123", 1234567890);
        gestionApostador.eliminarApostador(123456789);
        // Puedes agregar verificaciones adicionales aquí según la lógica de tu aplicación.
        // assertFalse("Condición esperada", condicionActual);
    }

    /**
     * Prueba para verificar la inicialización correcta de apostadores.
     */
    @Test
    public void testInicializarApostadores() {
        gestionApostador.inicializarApostadores();
        // Puedes agregar verificaciones adicionales aquí según la lógica de tu aplicación.
        // assertTrue("Condición esperada", condicionActual);
    }
}
